Remove init argument to push_clean_repos.
The init argument to push_clean_repos didn't really work so just run git init every time
to ensure the repository is created if it doesn't exist. Add a prefix argument instead
which allows, for example, a directory prefix to the output git repositories to make
an export with many repositories neater.